new document loader refactor to global util
